+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -6,8 +6,8 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.order.domain.OrderType;
 import kitchenpos.order.infra.KitchenridersClient;
 import org.springframework.stereotype.Service;
