+++ src/main/java/kitchenpos/ordertable/domain/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.ordertable.domain;
 
 import java.util.List;
 import java.util.Optional;
