+++ src/test/java/kitchenpos/tmp/application/InMemoryOrderRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.tmp.application;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 
 import java.util.*;
