+++ src/test/java/kitchenpos/tmp/application/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.tmp.application;
 
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 
 import java.util.*;
