+++ src/test/java/kitchenpos/tmp/application/OrderTableServiceTest.java
@@ -1,10 +1,10 @@
 package kitchenpos.tmp.application;
 
-import kitchenpos.order.application.OrderTableService;
+import kitchenpos.ordertable.application.OrderTableService;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.OrderTableRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
